moved finfo and iinfo methods to the Array API submodule, with the correct formatting.
