+++ ivy/functional/__init__.py
@@ -1,0 +1,2 @@
+from . import ivy
+from .ivy import *
