+++ ivy/functional/backends/numpy/__init__.py
@@ -1,6 +1,5 @@
 # global
 import sys
-import numpy as np
 
 # local
 from . import array_api
@@ -53,40 +52,6 @@ def closest_valid_dtype(type):
     if type_str in invalid_dtype_strs:
         return {'bfloat16': float16}[type_str]
     return type
-
-
-def iinfo(type):
-    return np.iinfo(dtype_from_str(type))
-
-
-class Finfo:
-
-    def __init__(self, np_finfo):
-        self._np_finfo = np_finfo
-
-    @property
-    def bits(self):
-        return self._np_finfo.bits
-
-    @property
-    def eps(self):
-        return float(self._np_finfo.eps)
-
-    @property
-    def max(self):
-        return float(self._np_finfo.max)
-
-    @property
-    def min(self):
-        return float(self._np_finfo.min)
-
-    @property
-    def smallest_normal(self):
-        return float(self._np_finfo.tiny)
-
-
-def finfo(type):
-    return Finfo(np.finfo(dtype_from_str(type)))
 
 
 backend = 'numpy'
