+++ ivy/functional/backends/tensorflow/__init__.py
@@ -1,7 +1,7 @@
 # global
 import sys
-import tensorflow as tf
-from tensorflow.python.framework.dtypes import DType
+from tensorflow.python.types.core import Tensor
+
 
 # local
 from . import array_api
@@ -52,40 +52,6 @@ def closest_valid_dtype(type):
     if type is None:
         return ivy.default_dtype()
     return type
-
-
-def iinfo(type):
-    return tf.experimental.numpy.iinfo(dtype_to_str(type))
-
-
-class Finfo:
-
-    def __init__(self, tf_finfo):
-        self._tf_finfo = tf_finfo
-
-    @property
-    def bits(self):
-        return self._tf_finfo.bits
-
-    @property
-    def eps(self):
-        return float(self._tf_finfo.eps)
-
-    @property
-    def max(self):
-        return float(self._tf_finfo.max)
-
-    @property
-    def min(self):
-        return float(self._tf_finfo.min)
-
-    @property
-    def smallest_normal(self):
-        return float(self._tf_finfo.tiny)
-
-
-def finfo(type):
-    return Finfo(tf.experimental.numpy.finfo(dtype_from_str(type)))
 
 
 backend = 'tensorflow'
