+++ ivy/functional/backends/torch/__init__.py
@@ -55,38 +55,4 @@ def closest_valid_dtype(type):
     return type
 
 
-def iinfo(type):
-    return _torch.iinfo(dtype_from_str(type))
-
-
-class Finfo:
-
-    def __init__(self, torch_finfo):
-        self._torch_finfo = torch_finfo
-
-    @property
-    def bits(self):
-        return self._torch_finfo.bits
-
-    @property
-    def eps(self):
-        return self._torch_finfo.eps
-
-    @property
-    def max(self):
-        return self._torch_finfo.max
-
-    @property
-    def min(self):
-        return self._torch_finfo.min
-
-    @property
-    def smallest_normal(self):
-        return self._torch_finfo.tiny
-
-
-def finfo(type):
-    return Finfo(_torch.finfo(dtype_from_str(type)))
-
-
 backend = 'torch'
